Update app icon and improve UI
Updated the path to the application's icon in build.gradle and added the icon to the main stage in DialogExtractor.java. The User Interface has also been improved, with changes including font adjustments and formatting in DialogExtractor.fxml, and addition of java.desktop as a required module in module-info.java.