Add detailed comments to test methods
In this commit, detailed comments have been added to various test methods from the LogUtilsTest and MinecraftLogTest classes. Each comment provides a brief overview of what the test case is testing, such as method functionality, the handling of different inputs, and expectations for exceptions. This enhances code readability and understanding.

Signed-off-by: Lilith Camplin <lilith.camplin@gmail.com>