Added file encoding selection to DiaLogExtractor
In this update, a ComboBox has been added to the DiaLogExtractor UI to let the user select the file encoding when processing a file. In addition, the processFile and decompressGzip methods were updated to support the provided encoding. Also, unnecessary individual javafx.* imports in DiaLogExtractor.fxml were replaced with general imports.