Add feature to auto-generate extracted dialog filename
The code was updated to include a feature where the filename of the extracted dialog is auto-generated based on the original file's name. This was implemented by storing the original filename and appending "_dialogExtracted.txt" to it during the file saving process. The initial filename is now also a parameter in the chooseFile function.

Signed-off-by: Lilith Camplin <lilith.camplin@gmail.com>