Update GitHub workflow to correctly name asset
In the GitHub release workflow file, the asset upload step has been modified. A run command is added to extract the base file name from the asset path and store it as ASSET_NAME. This change ensures that the asset is uploaded with the proper file name instead of its full path.

Signed-off-by: Lilith Camplin <lilith.camplin@gmail.com>